Refactor: Extract server action into a separate function
